requirements.txt
qiskit-terra
numpy
matplotlib
pillow